+++ pymc3/distributions/distribution.py
@@ -688,9 +688,10 @@ def draw_values(params, point=None, size=None):
         while stack:
             next_ = stack.pop(0)
             if (next_, size) in drawn:
-                # If the node already has a givens value, skip it
-                continue
-            elif isinstance(next_, (theano_constant, tt.sharedvar.SharedVariable)):
+                # If the node already has a givens value, add it to givens
+                givens[next_.name] = (next_, drawn[(next_, size)])
+            elif isinstance(next_, (theano_constant,
+                                    tt.sharedvar.SharedVariable)):
                 # If the node is a theano.tensor.TensorConstant or a
                 # theano.tensor.sharedvar.SharedVariable, its value will be
                 # available automatically in _compile_theano_function so
@@ -729,13 +730,9 @@ def draw_values(params, point=None, size=None):
                     # The node failed, so we must add the node's parents to
                     # the stack of nodes to try to draw from. We exclude the
                     # nodes in the `params` list.
-                    stack.extend(
-                        [
-                            node
-                            for node in named_nodes_descendents[next_]
-                            if node is not None and (node, size) not in drawn
-                        ]
-                    )
+                    stack.extend([node for node in named_nodes_parents[next_]
+                                  if node is not None and
+                                  getattr(node, "name", None) not in givens])
 
         # the below makes sure the graph is evaluated in order
         # test_distributions_random::TestDrawValues::test_draw_order fails without it
@@ -757,15 +754,10 @@ def draw_values(params, point=None, size=None):
                     evaluated[param_idx] = drawn[(param, size)]
                 else:
                     try:  # might evaluate in a bad order,
-                        # Sometimes _draw_value recurrently calls draw_values.
-                        # This may set values for certain nodes in the drawn
-                        # dictionary, but they don't get added to the givens
-                        # dictionary. Here, we try to fix that.
-                        if param in named_nodes_ancestors:
-                            for node in named_nodes_ancestors[param]:
-                                if node.name not in givens and (node, size) in drawn:
-                                    givens[node.name] = (node, drawn[(node, size)])
-                        value = _draw_value(param, point=point, givens=givens.values(), size=size)
+                        value = _draw_value(param,
+                                            point=point,
+                                            givens=givens.values(),
+                                            size=size)
                         evaluated[param_idx] = drawn[(param, size)] = value
                         givens[param.name] = (param, value)
                     except theano.gof.fg.MissingInputError: